+++ CHANGELOG.md
@@ -17,4 +17,5 @@ Follows the advice from [Keep A Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ### Known Issues
 * Filter entries that specify an ipset that don't exist does not raise an error
+* SNMPv3 Auth protocol of AES is not supported